+++ CHANGELOG.md
@@ -25,6 +25,9 @@ in 8.x, 7.x, 6.x, 5.x, 4.x, 3.x, 2.x, 1.x and 0.x versions.
 
 ## Lab
 
+### 2026-07-06
+ * Updated PackingSolver dependency
+ * Added `BinPackingCuttingCost` objective for `RectangleGuillotine` problem types
 ### 2026-07-05
  * Updated PackingSolver dependency
  * Fixed issue [#956](https://github.com/lairdubois/lairdubois-opencutlist-sketchup-extension/issues/956)